+++ src/Middleware/ThumbnailMiddleware.php
@@ -15,14 +15,12 @@ declare(strict_types=1);
  */
 namespace Thumber\Cake\Middleware;
 
-use Cake\Core\Configure;
 use Cake\Http\Response;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
 use Thumber\Cake\Http\Exception\ThumbNotFoundException;
-use Tools\Exceptionist;
 use Tools\Filesystem;
 
 /**
@@ -41,15 +39,13 @@ class ThumbnailMiddleware implements MiddlewareInterface
     {
         /** @var \Cake\Http\ServerRequest $request */
         $file = Filesystem::instance()->concatenate(THUMBER_TARGET, base64_decode($request->getParam('basename')));
-        Exceptionist::isReadable($file, __d('thumber', 'File `{0}` doesn\'t exist', $file), ThumbNotFoundException::class);
+        if (!is_readable($file)) {
+            throw new ThumbNotFoundException(__d('thumber', "File `{0}` doesn't exist", $file));
+        }
 
         $response = new Response();
         $response = $response->withModified(filemtime($file) ?: 0);
-        /**
-         * @todo to be removed in a later version
-         */
-        $method = version_compare(Configure::version(), '4.4', '>=') ? 'isNotModified' : 'checkNotModified';
-        if ($response->$method($request)) {
+        if ($response->isNotModified($request)) {
             return $response->withNotModified();
         }