+++ tests/TestCase/Command/ClearAllCommandTest.php
@@ -1,4 +1,5 @@
 <?php
+/** @noinspection PhpUnhandledExceptionInspection */
 declare(strict_types=1);
 
 /**
@@ -18,64 +19,40 @@ use Cake\Console\ConsoleIo;
 use Cake\Console\Exception\StopException;
 use Cake\Console\TestSuite\StubConsoleOutput;
 use Exception;
-use MeTools\TestSuite\ConsoleIntegrationTestTrait;
-use Thumber\Cake\TestSuite\TestCase;
+use MeTools\TestSuite\CommandTestCase;
+use Thumber\Cake\TestSuite\TestTrait;
 use Thumber\Cake\Utility\ThumbManager;
 
 /**
  * ClearAllCommandTest class
  * @property \Thumber\Cake\Command\ClearAllCommand $Command
  */
-class ClearAllCommandTest extends TestCase
+class ClearAllCommandTest extends CommandTestCase
 {
-    use ConsoleIntegrationTestTrait;
+    use TestTrait;
 
     /**
-     * @var bool
-     */
-    protected $autoInitializeClass = true;
-
-    /**
-     * @var string
-     */
-    protected $command = 'thumber.clear_all -v';
-
-    /**
-     * Tests for `execute()` method
      * @test
+     * @uses \Thumber\Cake\Command\ClearAllCommand::execute()
      */
     public function testExecute(): void
     {
+        $command = 'thumber.clear_all -v';
+
         $this->createSomeThumbs();
-        $this->exec($this->command);
-        $this->assertExitWithSuccess();
+        $this->exec($command);
+        $this->assertExitSuccess();
         $this->assertOutputRegExp('/^Thumbnails deleted: [^0]\d*$/');
-    }
 
-    /**
-     * Tests for `execute()` method, with no thumbnails
-     * @test
-     */
-    public function testExecuteNoThumbs(): void
-    {
-        $this->exec($this->command);
-        $this->assertExitWithSuccess();
+        //With no thumbnails
+        $this->exec($command);
+        $this->assertExitSuccess();
         $this->assertOutputContains('Thumbnails deleted: 0');
-    }
 
-    /**
-     * Tests for `execute()` method, on failure
-     * @test
-     */
-    public function testExecuteOnFailure(): void
-    {
+        //On failure
         $this->expectException(StopException::class);
-        $this->Command->ThumbManager = $this->getMockBuilder(ThumbManager::class)
-            ->onlyMethods(['_clear'])
-            ->getMock();
-        $this->Command->ThumbManager->method('_clear')
-            ->will($this->throwException(new Exception()));
-
+        $this->Command->ThumbManager = $this->createPartialMock(ThumbManager::class, ['_clear']);
+        $this->Command->ThumbManager->method('_clear')->willThrowException(new Exception());
         $this->Command->run([], new ConsoleIo(null, new StubConsoleOutput()));
     }
 }