+++ tests/TestCase/Utility/ThumbCreatorTest.php
@@ -1,4 +1,5 @@
 <?php
+/** @noinspection PhpUnhandledExceptionInspection */
 declare(strict_types=1);
 
 /**
@@ -14,80 +15,82 @@ declare(strict_types=1);
  */
 namespace Thumber\Cake\Test\TestCase\Utility;
 
-use BadMethodCallException;
 use Cake\Core\Configure;
+use ErrorException;
 use Intervention\Image\ImageManager;
-use InvalidArgumentException;
+use MeTools\Core\Plugin;
 use Thumber\Cake\TestSuite\TestCase;
 use Thumber\Exception\NotReadableImageException;
 use Thumber\Exception\UnsupportedImageTypeException;
 use Tools\Exception\NotReadableException;
 use Tools\Exception\NotWritableException;
+use Tools\TestSuite\ReflectionTrait;
 
 /**
  * ThumbCreatorTest class
  */
 class ThumbCreatorTest extends TestCase
 {
+    use ReflectionTrait;
+
     /**
      * Test for `__construct()` method, passing a no existing file
      * @test
+     * @uses \Thumber\Cake\Utility\ThumbCreator::__construct()
      */
     public function testConstructNoExistingFile(): void
     {
         $this->expectException(NotReadableException::class);
-        $this->expectExceptionMessage('File or directory `tests' . DS . 'test_app' . DS . 'webroot' . DS . 'img' . DS . 'noExistingFile.gif` does not exist');
+        $this->expectExceptionMessage('File or directory `' . WWW_ROOT . 'img' . DS . 'noExistingFile.gif` is not readable');
         $this->getThumbCreatorInstance('noExistingFile.gif');
     }
 
     /**
      * Test for `__construct()` method, passing a no existing file from plugin
      * @test
+     * @uses \Thumber\Cake\Utility\ThumbCreator::__construct()
      */
     public function testConstructNoExistingFileFromPlugin(): void
     {
-        $this->loadPlugins(['TestPlugin' => []]);
         $this->expectException(NotReadableException::class);
-        $this->expectExceptionMessage('File or directory `tests' . DS . 'test_app' . DS . 'Plugin' . DS . 'TestPlugin' . DS . 'webroot' . DS . 'img' . DS . 'noExistingFile.gif` does not exist');
+        $this->loadPlugins(['TestPlugin' => []]);
+        $this->expectExceptionMessage('File or directory `' . Plugin::path('TestPlugin') . 'webroot' . DS . 'img' . DS . 'noExistingFile.gif` is not readable');
         $this->getThumbCreatorInstance('TestPlugin.noExistingFile.gif');
     }
 
     /**
      * Test for `getImageInstance()` method, with unsupported image type for GD driver
-     * @ŧest
+     * @test
+     * @uses \Thumber\Cake\Utility\ThumbCreator::getImageInstance()
      */
     public function testGetImageInstanceUnsupportedImageType(): void
     {
         $this->expectException(UnsupportedImageTypeException::class);
         $this->expectExceptionMessage('Image type `image/jpeg` is not supported by this driver');
-        $thumbCreator = $this->getThumbCreatorInstance();
-        $thumbCreator->ImageManager = $this->getMockBuilder(ImageManager::class)
-            ->setMethods(['make'])
-            ->getMock();
-        $thumbCreator->ImageManager->method('make')->will($this->throwException(new UnsupportedImageTypeException('', 0, E_ERROR, __FILE__, __LINE__, null, 'image/jpeg')));
-        $this->invokeMethod($thumbCreator, 'getImageInstance');
+        $ThumbCreator = $this->getThumbCreatorInstance();
+        $ThumbCreator->ImageManager = $this->createPartialMock(ImageManager::class, ['make']);
+        $ThumbCreator->ImageManager->method('make')->willThrowException(new UnsupportedImageTypeException('Image type `image/jpeg` is not supported by this driver'));
+        $this->invokeMethod($ThumbCreator, 'getImageInstance');
     }
 
     /**
      * Test for `getImageInstance()` method, with a not readable image
-     * @ŧest
+     * @test
+     * @uses \Thumber\Cake\Utility\ThumbCreator::getImageInstance()
      */
     public function testGetImageInstanceNotReadableImageException(): void
     {
         $this->expectException(NotReadableImageException::class);
         $this->expectExceptionMessage('Unable to read image from file `anExampleFile`');
-
-        $thumbCreator = $this->getThumbCreatorInstance();
-        $thumbCreator->ImageManager = $this->getMockBuilder(ImageManager::class)
-            ->setMethods(['make'])
-            ->getMock();
-        $thumbCreator->ImageManager->method('make')->will($this->throwException(new NotReadableImageException('', 0, E_ERROR, __FILE__, __LINE__, null, 'anExampleFile')));
-        $this->invokeMethod($thumbCreator, 'getImageInstance');
+        $ThumbCreator = $this->getThumbCreatorInstance();
+        $ThumbCreator->ImageManager = $this->createPartialMock(ImageManager::class, ['make']);
+        $ThumbCreator->ImageManager->method('make')->willThrowException(new NotReadableImageException('Unable to read image from file `anExampleFile`'));
+        $this->invokeMethod($ThumbCreator, 'getImageInstance');
     }
 
     /**
-     * Test for `getUrl()` method
-     * @ŧest
+     * @test
+     * @uses \Thumber\Cake\Utility\ThumbCreator::getUrl()
      */
     public function testGetUrl(): void
     {
@@ -101,25 +104,25 @@ class ThumbCreatorTest extends TestCase
         $this->assertTextStartsNotWith(Configure::read('App.fullBaseUrl'), $result);
 
         //Without the `$target` property
-        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Missing path of the generated thumbnail. Probably the `save()` method has not been invoked');
         $this->getThumbCreatorInstance()->getUrl();
     }
 
     /**
-     * Test for `save()` method
      * @test
+     * @uses \Thumber\Cake\Utility\ThumbCreator::save()
      */
     public function testSave(): void
     {
         //When unable to create the file
         $this->assertException(function () {
             $this->getThumbCreatorInstance()->save();
-        }, BadMethodCallException::class, 'No valid method called before the `save()` method');
+        }, ErrorException::class, 'No valid method called before the `save()` method');
 
         //Without a valid method called before
         $this->skipIf(IS_WIN);
         $this->assertException(function () {
             $this->getThumbCreatorInstance()->resize(200)->save(['target' => DS . 'noExisting']);
-        }, NotWritableException::class, 'Unable to create file ``');
+        }, NotWritableException::class, 'Unable to create file `' . DS . 'noExisting`');
     }
 }