+++ tests/test_app/config/routes.php
@@ -1,8 +1,7 @@
 <?php
 declare(strict_types=1);
 
-use Cake\Routing\Router;
-
-Router::scope('/', function ($routes) {
+/** @var \Cake\Routing\RouteBuilder $routes */
+$routes->scope('/', function ($routes) {
     $routes->loadPlugin('Thumber/Cake');
 });